requirements.txt
attrs==19.3.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
django-registration==3.1
django-registration-redux==2.7
django-schema-graph==1.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
freezegun==0.3.15
gunicorn==20.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
